bot/pyproject.toml
[project]
name = "slack-llm-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt",
    "openai",
    "requests",
    "beautifulsoup4",
    "markdownify",
    "pdfplumber",
    "python-docx",
    "pandas",
    "python-dotenv",
    "apscheduler>=3.10",
    "pytz",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["bot", "agent", "tools", "config", "store", "news_digest"]
